+++ src/index.ts
@@ -46,4 +46,5 @@ export * from "./SharedFiles.js";
 export * from "./Status.js";
 export * from "./StatsGraphs.js";
 export * from "./Transmission.js";
+export * from "./Update.js";
 export * from "./Uploads.js";